feat(cactus-web): requested changes
CACTUS-604
